requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-genai==0.2.2